requirements.txt
pyyaml
torch
rasa[full]==2.1.2
tensorflow-text==2.4.2
fastapi>=0.62.0
mongomock==3.21.0
pytest==6.1.2
pytest-html==3.0.0
pytest-cov==2.10.1
pytest-xdist==2.1.0
pytest-asyncio==0.14.0
responses==0.12.0
pandas==1.1.4
mongoengine==0.20.0
pyjwt==1.7.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
py3dns==3.2.1
black
uvicorn==0.12.3
validators
secure==0.2.1
password-strength==0.0.3.post2
loguru
nltk==3.5
transformers==4.0.1
smart-config==0.1.3
moto
locust
greenlet==0.4.17
python-docx==0.8.10
PyMuPDF==1.18.4
sentencepiece
tensorflow==2.3.2
numpy==1.19.5
quart
pyTelegramBotApi
